Added a new C-block-based convenience API
This is handy for people using Clang, or other compilers (are there
any?) that support blocks. It lets you create or apply deltas and
consume the output incrementally, i.e. writing it to a file or socket,
instead of buffering the entire output in memory.
(This is my own contribution, not something I got from somewhere else.)